% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_convert_to_raw.R
\name{.change_byte}
\alias{.change_byte}
\title{Change the leading byte of a hash digest by one bit}
\usage{
.change_byte(hash_digest)
}
\description{
Used to test tampering with hash digests when verifying logs
}
