% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils-logging.R
\name{.cli_progress_output}
\alias{.cli_progress_output}
\title{Write a temporary msg to stdout that will be removed upon progressbar
completion}
\usage{
.cli_progress_output(msg, ..., .envir = parent.frame())
}
\description{
Write a temporary msg to stdout that will be removed upon progressbar
completion
}
\details{
Thin wrapper around cli::cli_progress_output that also logs
explicitly. Notably, it will not replace the progressbar when displaying
}
