% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{.concatenate_hashes}
\alias{.concatenate_hashes}
\title{Convert and concatenate two values into a single raw bytes objec}
\usage{
.concatenate_hashes(a, b, bits_a = NULL, bits_b = NULL)
}
\description{
Convert and concatenate two values into a single raw bytes objec
}
