% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils-datetimes.R
\name{.dt_to_ms}
\alias{.dt_to_ms}
\title{Convert from a lubridate-style datetime to a NIST 'milliseconds since epoch' representation}
\usage{
.dt_to_ms(dt)
}
\description{
Convert from a lubridate-style datetime to a NIST 'milliseconds since epoch' representation
}
