% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils-options_accessors.R
\name{.get_beacon_defaults}
\alias{.get_beacon_defaults}
\title{Get the options that need to be set for a specific Beacon to work}
\usage{
.get_beacon_defaults(beacon_name)
}
\description{
Currently only used for readability in BeaconInterface; facilitates triaging
of option sources; user-passes options take priority (in the case of
experienced developers extending functionality, e.g.), then global options (
allowing workflows that use .Rprofile or other ways of setting global
options) and then falling back to package defaults.
}
\seealso{
\code{\link[=.get_beacon_json]{.get_beacon_json()}}
}
