% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils-options_accessors.R
\name{.get_beacon_json}
\alias{.get_beacon_json}
\title{Read the JSON file specifying default settings values for a specific Beacon}
\usage{
.get_beacon_json()
}
\description{
Read the JSON file specifying default settings values for a specific Beacon
}
\seealso{
\code{\link[=.get_beacon_defaults]{.get_beacon_defaults()}}
}
