% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils-options_accessors.R
\name{.get_option}
\alias{.get_option}
\title{Get the value of a \code{beacr}-specific global option}
\usage{
.get_option(name, default = NULL)
}
\description{
Getter that prepends my pkg namespace to global options. Allows
something like \code{.get_option('hash_length')} without colliding
with another option from another pkg
}
