% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_convert_to_raw.R
\name{.is_integer_like}
\alias{.is_integer_like}
\title{Check if a value can be represented as an integer without loss}
\usage{
.is_integer_like(x)
}
\arguments{
\item{x}{A value to check}
}
\value{
\code{TRUE} if \code{x} can be converted to an integer without changing its value
}
\description{
Check if a value can be represented as an integer without loss
}
\concept{conversionutils}
