% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils-datetimes.R
\name{.ms_to_dt}
\alias{.ms_to_dt}
\title{Convert from NIST 'milliseconds-since-epoch' to a workable datetime via lubridate}
\usage{
.ms_to_dt(ms)
}
\description{
Convert from NIST 'milliseconds-since-epoch' to a workable datetime via lubridate
}
