% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils-datetimes.R
\name{.posix_to_NIST_UNIX}
\alias{.posix_to_NIST_UNIX}
\title{Convert from R's POSIXct data type to the time encoding used by the
NISTBeacon, namely \strong{milliseconds} since epoch}
\usage{
.posix_to_NIST_UNIX(posix_time)
}
\description{
Convert from R's POSIXct data type to the time encoding used by the
NISTBeacon, namely \strong{milliseconds} since epoch
}
\seealso{
\link{.dt_to_ms}
}
