% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{.preregistration_exists}
\alias{.preregistration_exists}
\title{Verify if any preregistration information was provided
Currently used in get_seed and verify_log}
\usage{
.preregistration_exists(preregistration_identifier, preregistration_source)
}
\description{
Verify if any preregistration information was provided
Currently used in get_seed and verify_log
}
