% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils-debugging.R
\name{.printt}
\alias{.printt}
\title{Print out NIST UNIX timestamps (milliseconds since epoch) in a readable way}
\usage{
.printt(msPosixtime)
}
\description{
Print out NIST UNIX timestamps (milliseconds since epoch) in a readable way
}
