% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{.prompt_menu}
\alias{.prompt_menu}
\title{Thin wrapper around \code{menu} to allow mocking in tests as mocking \code{menu} is
unsupported`}
\usage{
.prompt_menu(choices, title)
}
\description{
Thin wrapper around \code{menu} to allow mocking in tests as mocking \code{menu} is
unsupported`
}
