% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils-state_management.R
\name{.reset_data_dir}
\alias{.reset_data_dir}
\title{Delete the data directory where seed log and debug log files are stored by
default}
\usage{
.reset_data_dir(yes_to_all = FALSE)
}
\description{
ONLY intended for development use, e.g. to test starting from a fresh
installation
}
\keyword{internal}
