% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{.seed_log_exists}
\alias{.seed_log_exists}
\title{Check for seed log in directory}
\usage{
.seed_log_exists(dir_ = getwd(), seed_log_name = "seed_log.json")
}
\description{
Check for seed log in directory
}
