% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils-options_accessors.R
\name{.set_option}
\alias{.set_option}
\title{Set the value of a \code{beacr}-specific global option}
\usage{
.set_option(...)
}
\description{
Setter that prepends my pkg namespace to global options. Allows
something like \code{.set_option(hash_length=512L)} without colliding
with another option from another pkg
}
