% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils-datetimes.R
\name{.timeStamp_to_unix_time}
\alias{.timeStamp_to_unix_time}
\title{Converts a string return value from an API that represents a timestamp
into a timestamp integer in the format expected by the NISTBeacon; namely,
milliseconds since epoch}
\usage{
.timeStamp_to_unix_time(timeStamp, round = FALSE)
}
\description{
Converts a string return value from an API that represents a timestamp
into a timestamp integer in the format expected by the NISTBeacon; namely,
milliseconds since epoch
}
