% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_convert_to_raw.R
\name{.to_raw}
\alias{.to_raw}
\title{Dispatcher for conversion to atomic raw byte vector type}
\usage{
.to_raw(value, bits)
}
\arguments{
\item{value}{The value to be converted}

\item{bits}{The number of bits the output should have. This is important for validation but also detecting which conversion type to use; a string of 0's and 1's could be representing either binary or hex, so knowing how many bits it's supposed to represent is important in that edge case.}
}
\description{
Handle multiple possible raw data representations and convert them to the base R type. Offers extensibility for Beacons or other randomness ingest sources that represent binary data in nonstandard ways.
}
\keyword{conversionutils}
