% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read_logs.R
\name{read_logs}
\alias{read_logs}
\title{Read the debug log}
\usage{
read_logs(log_file = NULL)
}
\arguments{
\item{log_file}{The custom log file path you wish to read. If
no file is passed, a prompt menu will appear asking which log in the
default log file directory you wish to read}
}
\description{
Opens a new R window that allows reading the log file using the built-in
\code{file.show()} R function, starting from the bottom (most recently logged
messages). If no log file path is passed, will search for log files in the
default logging directory
}
\seealso{
\code{\link[=file.show]{file.show()}}
}
