% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_swirl.R
\name{swirl}
\alias{swirl}
\title{Mix together the output of multiple randomness beacons to produce a random
number robust to protocol deviations from any one beacon}
\usage{
swirl(...)
}
\description{
Currently this is a placeholder; when called it will perform the exact same behavior as \link{get_seed}. In the future when multiple beacons are implemented, this function will combine multiple beacons together.
}
\seealso{
\link{get_seed}
}
