% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beacons_BeaconInterface.R
\name{BeaconInterface}
\alias{BeaconInterface}
\title{Public interface for implementing a new \code{beacr}-compatible Beacon object from
a R6 class.}
\description{
Public interface for implementing a new \code{beacr}-compatible Beacon object from
a R6 class.
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-BeaconInterface-initialize}{\code{BeaconInterface$new()}}
    \item \href{#method-BeaconInterface-replay_file_exists}{\code{BeaconInterface$replay_file_exists()}}
    \item \href{#method-BeaconInterface-resolve_hash_function}{\code{BeaconInterface$resolve_hash_function()}}
    \item \href{#method-BeaconInterface-is_valid_hash}{\code{BeaconInterface$is_valid_hash()}}
    \item \href{#method-BeaconInterface-get_pulse_by_index}{\code{BeaconInterface$get_pulse_by_index()}}
    \item \href{#method-BeaconInterface-get_description}{\code{BeaconInterface$get_description()}}
    \item \href{#method-BeaconInterface-get_latest_seed}{\code{BeaconInterface$get_latest_seed()}}
    \item \href{#method-BeaconInterface-get_latest_pulse}{\code{BeaconInterface$get_latest_pulse()}}
    \item \href{#method-BeaconInterface-get_pulse_index}{\code{BeaconInterface$get_pulse_index()}}
    \item \href{#method-BeaconInterface-get_pulse_time}{\code{BeaconInterface$get_pulse_time()}}
    \item \href{#method-BeaconInterface-validate_pulse}{\code{BeaconInterface$validate_pulse()}}
    \item \href{#method-BeaconInterface-hash}{\code{BeaconInterface$hash()}}
    \item \href{#method-BeaconInterface-clone}{\code{BeaconInterface$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BeaconInterface-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-BeaconInterface-initialize}{}}}
\subsection{\code{BeaconInterface$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BeaconInterface$new(...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BeaconInterface-replay_file_exists"></a>}}
\if{latex}{\out{\hypertarget{method-BeaconInterface-replay_file_exists}{}}}
\subsection{\code{BeaconInterface$replay_file_exists()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BeaconInterface$replay_file_exists()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BeaconInterface-resolve_hash_function"></a>}}
\if{latex}{\out{\hypertarget{method-BeaconInterface-resolve_hash_function}{}}}
\subsection{\code{BeaconInterface$resolve_hash_function()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BeaconInterface$resolve_hash_function()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BeaconInterface-is_valid_hash"></a>}}
\if{latex}{\out{\hypertarget{method-BeaconInterface-is_valid_hash}{}}}
\subsection{\code{BeaconInterface$is_valid_hash()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BeaconInterface$is_valid_hash(hashed_value, comparison_value = NULL)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BeaconInterface-get_pulse_by_index"></a>}}
\if{latex}{\out{\hypertarget{method-BeaconInterface-get_pulse_by_index}{}}}
\subsection{\code{BeaconInterface$get_pulse_by_index()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BeaconInterface$get_pulse_by_index(...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BeaconInterface-get_description"></a>}}
\if{latex}{\out{\hypertarget{method-BeaconInterface-get_description}{}}}
\subsection{\code{BeaconInterface$get_description()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BeaconInterface$get_description()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BeaconInterface-get_latest_seed"></a>}}
\if{latex}{\out{\hypertarget{method-BeaconInterface-get_latest_seed}{}}}
\subsection{\code{BeaconInterface$get_latest_seed()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BeaconInterface$get_latest_seed(...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BeaconInterface-get_latest_pulse"></a>}}
\if{latex}{\out{\hypertarget{method-BeaconInterface-get_latest_pulse}{}}}
\subsection{\code{BeaconInterface$get_latest_pulse()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BeaconInterface$get_latest_pulse(...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BeaconInterface-get_pulse_index"></a>}}
\if{latex}{\out{\hypertarget{method-BeaconInterface-get_pulse_index}{}}}
\subsection{\code{BeaconInterface$get_pulse_index()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BeaconInterface$get_pulse_index(...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BeaconInterface-get_pulse_time"></a>}}
\if{latex}{\out{\hypertarget{method-BeaconInterface-get_pulse_time}{}}}
\subsection{\code{BeaconInterface$get_pulse_time()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BeaconInterface$get_pulse_time(...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BeaconInterface-validate_pulse"></a>}}
\if{latex}{\out{\hypertarget{method-BeaconInterface-validate_pulse}{}}}
\subsection{\code{BeaconInterface$validate_pulse()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BeaconInterface$validate_pulse(...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BeaconInterface-hash"></a>}}
\if{latex}{\out{\hypertarget{method-BeaconInterface-hash}{}}}
\subsection{\code{BeaconInterface$hash()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BeaconInterface$hash(...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BeaconInterface-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BeaconInterface-clone}{}}}
\subsection{\code{BeaconInterface$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BeaconInterface$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
