% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read_logs.R
\name{read_logs}
\alias{read_logs}
\title{Read the debug log}
\usage{
read_logs(log_file = NULL)
}
\arguments{
\item{log_file}{The custom log file path you wish to read. If
no file is passed, a prompt menu will appear asking which log in the
default log file directory you wish to read}
}
\description{
Read the debug log
}
\seealso{
\code{\link[=file.show]{file.show()}}
}
