% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_verify_log.R
\name{verify_log}
\alias{verify_log}
\title{Verify the contents of a seed log are internally consistent}
\usage{
verify_log(log_file, start_pulse_for_skiplist = NULL)
}
\arguments{
\item{start_pulse_for_skiplist}{To allow starting from a diff pulse than
latest on chain, i.e. if a known diversion from protocol or infiltration
occurred}
}
\description{
Verify the contents of a seed log are internally consistent
}
\details{
Currently uses a hash table variable labelled \code{.beacons}
that matches a beacon string to the corresponding R6 object
}
